backend/pyproject.toml
[project]
name = "crawler-demo"
version = "0.1.0"
description = "A web crawler, scraper, and archiving service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.12.4",
    "litestar[sqlalchemy,standard]>=2.0.0",
    "celery[sqlalchemy]>=5.3.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.21.0",
    "crawlbase>=1.0.0",
    "markdownify>=1.1.0",
    "litellm>=1.74.15.post1",
    "python-dotenv>=1.1.1",
]
